+++ src/SexyAppFramework/sound/SoundManager.h
@@ -19,15 +19,15 @@ public:
 
 	virtual bool			Initialized() = 0;
 
-	virtual bool			LoadSound(unsigned int theSfxID, const std::string& theFilename) = 0;
-	virtual int				LoadSound(const std::string& theFilename) = 0;
-	virtual void			ReleaseSound(unsigned int theSfxID) = 0;
+	virtual bool			LoadSound(intptr_t theSfxID, const std::string& theFilename) = 0;
+	virtual intptr_t		LoadSound(const std::string& theFilename) = 0;
+	virtual void			ReleaseSound(intptr_t theSfxID) = 0;
 
 	virtual void			SetVolume(double theVolume) = 0;
-	virtual bool			SetBaseVolume(unsigned int theSfxID, double theBaseVolume) = 0;
-	virtual bool			SetBasePan(unsigned int theSfxID, int theBasePan) = 0;	
+	virtual bool			SetBaseVolume(intptr_t theSfxID, double theBaseVolume) = 0;
+	virtual bool			SetBasePan(intptr_t theSfxID, int theBasePan) = 0;
 
-	virtual SoundInstance*	GetSoundInstance(unsigned int theSfxID) = 0;
+	virtual SoundInstance*	GetSoundInstance(intptr_t theSfxID) = 0;
 
 	virtual void			ReleaseSounds() = 0;
 	virtual void			ReleaseChannels() = 0;
@@ -37,9 +37,8 @@ public:
 
 	virtual void			Flush() = 0;
 	virtual void			StopAllSounds() = 0;
-	virtual int				GetFreeSoundId() = 0;
+	virtual intptr_t		GetFreeSoundId() = 0;
 	virtual int				GetNumSounds() = 0;
-
 };
